CORE-1134 introduced Effect F in base ISpace trait
