+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/dispatch.scala
@@ -9,7 +9,7 @@ import coop.rchain.models.TaggedContinuation.TaggedCont.{Empty, ParBody, ScalaBo
 import coop.rchain.models._
 import coop.rchain.rholang.interpreter.accounting.{Cost, CostAccount, CostAccountingAlg}
 import coop.rchain.rholang.interpreter.storage.TuplespaceAlg
-import coop.rchain.rspace.ISpace
+import coop.rchain.rspace.{FreudianSpace, ISpace}
 import cats.implicits._
 import coop.rchain.rspace.pure.PureRSpace
 
@@ -63,11 +63,11 @@ class RholangOnlyDispatcher[M[_]] private (_reducer: => Reduce[M])(implicit s: S
 
 object RholangOnlyDispatcher {
 
-  def create[M[_], F[_]](tuplespace: ISpace[Channel,
-                                            BindPattern,
-                                            ListChannelWithRandom,
-                                            ListChannelWithRandom,
-                                            TaggedContinuation],
+  def create[M[_], F[_]](tuplespace: FreudianSpace[Channel,
+                                                   BindPattern,
+                                                   ListChannelWithRandom,
+                                                   ListChannelWithRandom,
+                                                   TaggedContinuation],
                          urnMap: Map[String, Par] = Map.empty)(
       implicit
       parallel: Parallel[M, F],
@@ -127,11 +127,11 @@ class RholangAndScalaDispatcher[M[_]] private (
 object RholangAndScalaDispatcher {
 
   def create[M[_], F[_]](
-      tuplespace: ISpace[Channel,
-                         BindPattern,
-                         ListChannelWithRandom,
-                         ListChannelWithRandom,
-                         TaggedContinuation],
+      tuplespace: FreudianSpace[Channel,
+                                BindPattern,
+                                ListChannelWithRandom,
+                                ListChannelWithRandom,
+                                TaggedContinuation],
       dispatchTable: => Map[Long, Function1[Seq[ListChannelWithRandom], M[Unit]]],
       urnMap: Map[String, Par])(
       implicit
