+++ rspace/src/main/scala/coop/rchain/rspace/RSpace.scala
@@ -21,7 +21,7 @@ class RSpace[C, P, A, R, K](store: IStore[C, P, A, K], branch: Branch)(
     serializeA: Serialize[A],
     serializeK: Serialize[K]
 ) extends RSpaceOps[C, P, A, R, K](store, branch)
-    with ISpace[C, P, A, R, K] {
+    with FreudianSpace[C, P, A, R, K] {
 
   override protected[this] val logger: Logger = Logger[this.type]
